requirements.txt
streamlit
google-generativeai
PyPDF2
python-dotenv
prebase4
setuptools
